refactor(upload): seperate file upload components
fileList / fileItem / trigger
