+++ shared/design/src/components/upload/index.tsx
@@ -3,15 +3,15 @@ import * as React from "react"
 import { cn } from "../../utils/cn"
 import { FileUploadDropzone } from "./components/dropzone"
 import {
-  FileUpload,
   FileUploadClear,
   FileUploadItem,
   FileUploadItemDelete,
   FileUploadItemMetadata,
   FileUploadItemProgress,
-  FileUploadList,
-  FileUploadTrigger,
-} from "./upload"
+} from "./components/fileItem"
+import { FileUploadList } from "./components/fileList"
+import { FileUploadTrigger } from "./components/trigger"
+import { FileUpload } from "./upload"
 
 function Dropzone({
   className,
@@ -30,25 +30,6 @@ function Dropzone({
     </FileUploadDropzone>
   )
 }
-
-// const TriggerContent = () => {
-//   return (
-//     <div className="flex items-center justify-center gap-[22px] text-text-interactive-tertiary">
-//       <div className="typography-heading-lg-semibold">파일 업로드</div>
-//       <div className="typography-heading-sm-medium">
-//         JPG, JPEG, PNG (최대 2MB)
-//       </div>
-//     </div>
-//   )
-// }
-
-// const TriggerButton = ({ children }) => {
-//   return (
-//     <PrimaryBoxButton size="md" _style="solid">
-//       {children}
-//     </PrimaryBoxButton>
-//   )
-// }
 
 function Trigger({
   className,
